Made resolving truly optional, added streaming results + misc
